pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SEMSEG"
version = "1.0.0"
description = "3D FIB-SEM Segmentation and Quantification Program"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Manus AI"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "scikit-image>=0.18.0",
    "h5py>=3.6.0",
    "tifffile>=2021.11.2",
    "PyYAML>=6.0",
    "Pillow>=9.0.0",
    "zarr",
    "s3fs",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
deep_learning = ["tensorflow>=2.8.0"]
advanced_segmentation = ["PyMaxflow>=1.2.13"]
all = ["tensorflow>=2.8.0", "PyMaxflow>=1.2.13"]

[project.scripts]
semseg = "core.cli:main"

[tool.setuptools]
packages = ["core", "pipeline", "gui"]
py-modules = ["__init__", "__main__", "create_sample_data", "demo_gui", "gui", "launch_gui", "neuroglancer_integration", "webui"]
